+++ packages/core-typings/src/IMethodType.ts
@@ -1,5 +1,0 @@
-import type { IMethodThisType } from './IMethodThisType';
-
-export interface IMethodType {
-	[key: string]: (this: IMethodThisType, ...args: any[]) => any;
-}